Additional accessibility + new images
The user can now tab through & select section headers, which lets screen readers know which section of the webpage is in the viewport. A tabindex of 0 was only applied to the smaller h2 headers as those are in English. Some background images were changed for greater legibility. Lastly, some gallery images in Geography.js were changed & more detailed description text was added.