AJAX live weather update
Live weather can now be viewed inCities.js for Istanbul, Ankara, Izmir, Bursa, & Adana. The weather comes from OpenWeatherMap's API & displays current temperature, cloud coverage, & wind. The weather is then inserted into additional li's that were added to all 5 sections of Cities.js. Also when the API returns 'clear sky', the if statement replaces the text with 'clear skies', which sounds more natural. Additionally, more in-depth text was added to the gallery in Geography.js, as well as a new background image for .Three and new background images for the chevrons. These images are not only better but also contrast more with the text, making it more legible.