Browser icon now live
Adjustments were made to the webpack regular expression & now it accepts .ico files. A red crescent & star is now displayed in the browser tab. Also for some reason the file extention for Izmir-3 was .JPG instead of .jpg, causing a webpack error. The capital file extention was changed to lower case & the file was renamed to Izmir-1.jpg, as there is only one photo of Izmir on the entire site.