Dynamic Sidebar
The sidebar icons now change color based on the color of the background div. For example, if the div in the background is dark, the icons will be white. If the div in the background is white (or a light color), the icons will turn black. Also, the icons were changed to ones that are easier to see. Lastly, the scss was adjusted for the sidebar & an Icons8 link was added to the bottom navigation.