Improved semantics + smaller img files
The former numerical naming system for large divs has been replaced with a more logical nomenclature based on its content. Also, file sizes have been reduced dramatically, with the img folder being more than 50% smaller. Lastly, the icons in the Sidebar are now white with a black drop-shadow. Not only does this fit the color scheme of the website better, but it also improves visibility so that the icons don't have to toggle between two colors like they did before.