Small tweaks
A new favicon.ico has been added to the page & the background image of .Cities-3 was changed. The new background image is an evening image of Izmir, which improves the legibility of the text. Also, the div height of .Intro-1 has been changed to 100vh & the opacity of the white sidebar icons has been increased slightly from 0.6 to 0.7. Lastly the parallax effect on .Geography-2 has been reduced so it's less jarring for the user.