More API/React use & new pronunciation audio
New API calls are displayed in Cities.js which show the longitude/latitude of the cities' locations. In Languages.js, the user can now hear verbal pronunciations of basic Turkish phrases by clicking the audio symbol next to the word. Also, all classNames with capital letters have been replaced with nomenclature that only uses lower case letters. Lastly, a div in Hotels.js is in the middle of a revamp. A search function is planned that will allow the user to search for a hotel in Turkey using a seachbox & API.