Active sidebar
When the user scrolls to a certain section of the page, the corresponding sidebar element grows slightly in size & moves to the left. When the user scrolls out of the section, the sidebar sidebar element shrinks & moves back into place. The opacity of sidebar icons has been adjusted & the active sidebar icon now becomes opaque. Lastly, .intro-1-div now has a new background image that makes the text more legible. The headers #hos-geldiniz & #welcome no longer neeed a text shadow to improve visibility & are now aligned to the left of the div. Google maps integration is coming in the next update.